# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Difficulty level
 */
export type GenerateQuizRequestDifficulty =
  (typeof GenerateQuizRequestDifficulty)[keyof typeof GenerateQuizRequestDifficulty];

export const GenerateQuizRequestDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface GenerateQuizRequest {
  /** The quiz topic (e.g. Math, Science, Animals) */
  topic: string;
  /** Difficulty level */
  difficulty: GenerateQuizRequestDifficulty;
}

export interface QuizQuestion {
  question: string;
  /**
   * @minItems 4
   * @maxItems 4
   */
  options: string[];
  /**
   * @minimum 0
   * @maximum 3
   */
  correct: number;
  /** A helpful hint to guide the student without giving away the answer */
  hint: string;
  explanation: string;
  fun_fact: string;
}

export interface GenerateQuizResponse {
  source_summary: string;
  questions: QuizQuestion[];
}

export interface ErrorResponse {
  error: string;
}