# lib/api-zod/src/generated/types/quizQuestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuizQuestion {
  question: string;
  /**
   * @minItems 4
   * @maxItems 4
   */
  options: string[];
  /**
   * @minimum 0
   * @maximum 3
   */
  correct: number;
  /** A helpful hint to guide the student without giving away the answer */
  hint: string;
  explanation: string;
  fun_fact: string;
}